chip8_emulator::decoder::instruction

Enum Instruction

pub enum Instruction {

    CallMachineCode {
        address: Address,
    },
    ClearDisplay,
    Return,
    Goto {
        address: Address,
    },
    CallSubroutine {
        address: Address,
    },
    EqConst {
        x: Index,
        value: u8,
    },
    NeqConst {
        x: Index,
        value: u8,
    },
    EqReg {
        x: Index,
        y: Index,
    },
    AssignConst {
        x: Index,
        value: u8,
    },
    AddAssignConst {
        x: Index,
        value: u8,
    },
    AssignReg {
        x: Index,
        y: Index,
    },
    OrReg {
        x: Index,
        y: Index,
    },
    AndReg {
        x: Index,
        y: Index,
    },
    XorReg {
        x: Index,
        y: Index,
    },
    AddAssignReg {
        x: Index,
        y: Index,
    },
    SubAssignReg {
        x: Index,
        y: Index,
    },
    RShift {
        x: Index,
        y: Index,
    },
    SubAssignRegInverse {
        x: Index,
        y: Index,
    },
    LShift {
        x: Index,
        y: Index,
    },
    NeqReg {
        x: Index,
        y: Index,
    },
    SetI {
        address: Address,
    },
    GotoPlusV0 {
        address: Address,
    },
    Rand {
        x: Index,
        value: u8,
    },
    DrawSprite {
        x: Index,
        y: Index,
        height: SpriteHeight,
    },
    KeyPressedSkip {
        x: Index,
    },
    KeyReleasedSkip {
        x: Index,
    },
    GetDelayTimer {
        x: Index,
    },
    AwaitKeyPress {
        x: Index,
    },
    SetDelayTimer {
        x: Index,
    },
    SetSoundTimer {
        x: Index,
    },
    AddAssignAddress {
        x: Index,
    },
    SetSpriteAddr {
        x: Index,
    },
    SetBCD {
        x: Index,
    },
    DumpRegisters {
        x: Index,
    },
    LoadRegisters {
        x: Index,
    },

}

All the opcodes for chip8 emulator

NNN - 12-bit address NN - 8-bit constant N - 4-bit constant X/Y - 4-bit register code (0-F)

Variants

CallMachineCode

0NNN, call machine code routine

Fields

address: Address

Address of the subroutine

ClearDisplay

00E0, clears the display

Return

00EE, returns from the function

Goto

1NNN, jump to address NNN

Fields

address: Address

address to jump to

CallSubroutine

2NNN, call subroutine at address NNN

Fields

address: Address

subroutine start address

EqConst

3XNN, if VX == NN, skip next instruction

Fields

x: Index

register index for VX

value: u8

value to compare to

NeqConst

4XNN, if VX != NN, skip next instruction

Fields

x: Index

register index for VX

value: u8

value to compare to

EqReg

5XY0, if VX == VY, skip next instruction

Fields

x: Index

register index for VX

y: Index

register index for VY

AssignConst

6XNN, set VX to NN

Fields

x: Index

register index for VX

value: u8

value to assign

AddAssignConst

7XNN, VX += NN

Fields

x: Index

register index for VX

value: u8

value to assign

AssignReg

8XY0, set VX to VY

Fields

x: Index

register index for VX

y: Index

register index for VY

OrReg

8XY1, VX |= VY

Fields

x: Index

register index for VX

y: Index

register index for VY

AndReg

8XY2, VX &= VY

Fields

x: Index

register index for VX

y: Index

register index for VY

XorReg

8XY3, VX ^= VY

Fields

x: Index

register index for VX

y: Index

register index for VY

AddAssignReg

8XY4, VX += VY

Fields

x: Index

register index for VX

y: Index

register index for VY

SubAssignReg

8XY5, VX -= VY

Fields

x: Index

register index for VX

y: Index

register index for VY

RShift

8XY6, VX >>= 1

Fields

x: Index

register index for VX

y: Index

register index for VY

SubAssignRegInverse

8XY7, VX = VY - VX

Fields

x: Index

register index for VX

y: Index

register index for VY

LShift

8XYE, VX <<= 1

Fields

x: Index

register index for VX

y: Index

register index for VY

NeqReg

9XY0, if VX != VY, skip next instruction

Fields

x: Index

register index for VX

y: Index

register index for VY

SetI

ANNN, set I to NNN

Fields

address: Address

address to set I to

GotoPlusV0

BNNN, jump to address NNN + V0

Fields

address: Address

address to jump to

Rand

CXNN, VX = rand() & NN

Fields

x: Index

register index for VX

value: u8

constant to and with

DrawSprite

DXYN, draw sprite at VX, VY with height being N

Fields

x: Index

register index for VX

y: Index

register index for VY

height: SpriteHeight

sprite height

KeyPressedSkip

EX9E, skip next instruction if key pressed matches with one stored in VX

Fields

x: Index

register index for VX

KeyReleasedSkip

EXA1, skip next instruction if key pressed doesn’t match with one stored in VX

Fields

x: Index

register index for VX

GetDelayTimer

FX07, set VX to delay timer value

Fields

x: Index

register index for VX

AwaitKeyPress

FX0A, wait until any key is pressed and store it to VX

Fields

x: Index

register index for VX

SetDelayTimer

FX15, set delay timer to VX

Fields

x: Index

register index for VX

SetSoundTimer

FX18, set sound timer to VX

Fields

x: Index

register index for VX

AddAssignAddress

FX1E, I += VX

Fields

x: Index

register index for VX

SetSpriteAddr

FX29, I = sprite_addr[VX]

Fields

x: Index

register index for VX

SetBCD

FX33, store BCD of VX to I

Fields

x: Index

register index for VX

DumpRegisters

FX55, dump register V0..=VX in memory starting at I

Fields

x: Index

register index for VX

LoadRegisters

FX65, load register V0..=VX from memory starting at I

Fields

x: Index

register index for VX