chip8_emulator::machine::cpu

Struct Cpu

pub struct Cpu {
    general: [u8; 16],
    address: u16,
    delay_timer: u8,
    sound_timer: u8,
    program_counter: u16,
    stack_pointer: usize,
    stack: [u16; 16],
    pub(crate) random_engine: SmallRng,
}

Store register and stack state\

Fields
general: [u8; 16]

General purpose registers, V0 through VF, 8 bit each

address: u16

Address register

delay_timer: u8

Delay timer

sound_timer: u8

Sound timer

program_counter: u16

Program counter

stack_pointer: usize

Stack pointer

stack: [u16; 16]

Stack

random_engine: SmallRng

Random engine for reproducible randomness

Implementations

impl Cpu

pub fn new() -> Self

Create a new CPU

pub fn program_counter(&self) -> u16

Get program counter

pub fn advance_program_counter(&mut self, step: u16) -> Result<(), CpuError>

Advance program counter

Returns an error if the program counter doesnโt fit into 12 bits

pub fn set_program_counter(&mut self, new_pc: u16) -> Result<(), CpuError>

Set program counter

Returns an error if the program counter doesnโt fit into 12 bits

pub fn address(&self) -> u16

Get address register (I)

pub fn set_address(&mut self, value: u16) -> Result<(), CpuError>

Set a new address register value (I)

Returns an error if the address doesnโt fit into 12 bits

pub fn advance_address(&mut self, step: u16) -> Result<(), CpuError>

Increment address register by given step

Returns an error if the address doesnโt fit into 12 bits

pub fn stack_push(&mut self, address: u16) -> Result<(), CpuError>

Push an addres onto stack, updating stack pointer accordingly

Returns an error if the stack limit is reached

pub fn stack_pop(&mut self) -> Result<u16, CpuError>

Pop an address from the stack, returning it

Returns an error if was called on an empty stack

pub fn vx(&mut self, x: Index) -> &mut u8

Helper to get the VX register, reduced boilerplate

pub fn delay_timer(&self) -> u8

Get delay timer value

pub fn sound_timer(&self) -> u8

Get sound timer value

pub fn set_delay_timer(&mut self, value: u8)

Set delay timer

pub fn set_sound_timer(&mut self, value: u8)

Set sound timer

pub fn random(&mut self) -> u8

Get randomness

pub fn tick_timers(&mut self)

Tick timers down by one if possible