chip8_emulator

Module window

Module that contains the window logic

Structs

App 🔒

The main emulator application

Constants

CPU_CYCLE_INTERVAL 🔒

The time interval for 500hz (simulate chip8 cpu at 500hz)

TIMER_INTERVAL 🔒

The time interval for 60hz (timers for chip8 operate on 60hz)

Functions

keymap 🔒

Map the real input to hex keyboard of chip8

load_program 🔒

Load the program from path and return a ready chip8 instance

run_app

Runs the main application of the emulator