chip8_emulator::window

Struct App

struct App<'a> {
    window: Option<Arc<Window>>,
    window_id: Option<WindowId>,
    pixels: Option<Pixels<'a>>,
    chip8: Chip8,
    last_ticked: Instant,
    last_cpu_cycle: Instant,
    prev_display_hash: Option<u64>,
    stream_handle: Option<OutputStream>,
    sound_sink: Option<Arc<Mutex<Sink>>>,
}

The main emulator application

Fields
window: Option<Arc<Window>>

Application’s window

window_id: Option<WindowId>

Window’s ID

pixels: Option<Pixels<'a>>

Pixels struct to draw on the window

chip8: Chip8

Chip8 instance

last_ticked: Instant

Last time the timers were ticked down

last_cpu_cycle: Instant

Last time the cpu cycle happened

prev_display_hash: Option<u64>

Hash of the last frame to consider rerendering

stream_handle: Option<OutputStream>

Stream handle for audio

sound_sink: Option<Arc<Mutex<Sink>>>

Audio sink

Implementations

impl<'a> App<'a>

fn from_chip8(chip8: Chip8) -> Self

Create an application struct from a ready chip8 instance

impl<'a> App<'a>

const DECAY: f32 = 0.25f32

Decay factor for the phosphor persitence

fn maybe_redraw_display(&mut self)

Determine if the current display is different from the last, and if yes, rerender

fn draw_display(&mut self, display_state: [[bool; 64]; 32])

Redraw the display depending on the current chip8 display state

fn tick_timers_and_sound(&mut self, now: Instant)

Tick the cpu timers and play sound if needed

fn play_sound(&self)

Play the sound

fn pause_sound(&self)

Pause the current sound

fn run_cpu_cycle(&mut self, now: Instant)

Run one cpu cycle